getPourcentage pour fidelite et premiere attraction

+++ Modele/Reduction.java
@@ -3,10 +3,10 @@ package Modele;
 public class Reduction {
     private int id_reduction;
     private String nom;
-    private double pourcentage;
+    private int pourcentage;
     private String description;
 
-    public Reduction(int id_reduction, String nom, double pourcentage, String description){
+    public Reduction(int id_reduction, String nom, int pourcentage, String description){
         this.id_reduction=id_reduction;
         this.nom = nom;
         this.pourcentage=pourcentage;
@@ -15,7 +15,7 @@ public class Reduction {
 
     public int getId_reduction(){return id_reduction;}
     public String getNom(){return nom;}
-    public double getPourcentage(){return pourcentage;}
+    public int getPourcentage(){return pourcentage;}
     public String getDescription(){return description;}
 
 }
